refactor module names and ui
